Add script to clone project repos

+++ windows-setup/setup.ps1
@@ -11,6 +11,7 @@
 [console]::beep(180,500)  # Audio signal for input
 $user = read-host -p user
 
+<#
 echo ''
 echo '__________________________________________________'
 echo 'Setup vim backup directories'
@@ -27,43 +28,39 @@ $env:SCOOP="$HOME/apps/scoop"
 Set-ExecutionPolicy RemoteSigned -scope CurrentUser -Force -Confirm:$false
 [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
 iwr get.scoop.sh -UseBasicParsing | iex
-
+#>
 echo ''
 echo '__________________________________________________'
-echo 'Install git, gh, and configure user, and decrypt .ssh'
-scoop install git gh
+echo 'Install git and configure user, and decrypt .ssh'
+scoop install git
 git config --global user.name $user
 git config --global user.email $user@gmail.com
 
 echo ''
 echo '__________________________________________________'
-echo 'Get USB drive letter'
-# Get the drive letter since this can change from machine to machine
+echo 'Decrypt ssh keys then delete gnupg directory'
 $usb = (Get-WmiObject -Class Win32_LogicalDisk | where DriveType -eq 2).DeviceID
 cd $usb
-
-echo ''
-echo '__________________________________________________'
-echo 'Decrypt ssh keys then delete gnupg directory'
-[console]::beep(180,500)  # Audio signal for input
-git-bash -c "cd $usb ; gpg -d .ssh | tar -xzC ~/ ; rm -r ~/.gnupg"
+git-bash -c "echo -en '\007'; cd $usb ; pwd ; gpg -d keys | tar -xzC ~/ ; rm -r ~/.gnupg"
+if ( -not (Test-Path -Path '~/.ssh' -PathType Container) ) { [Environment]::SetEnvironmentVariable('SCOOP', $env:SCOOP, 'User') ; read-host ABORT: Keys not loaded ; exit }
 # to compress and encrypt
-#   tar -czf .ssh | gpg -c > .ssh~ ; mv -f .ssh~ .ssh
+#   tar -cz .ssh | gpg -c > keys ; rm -r .ssh
 # then delete .ssh and rename .ssh~ to .ssh
 mv -force '~/.ssh/Github CLI' ~/AppData/Roaming/
 
 echo ''
 echo '__________________________________________________'
-echo 'Clone projects from git'
+echo 'Clone config files'
 #	Git needs to clone to an empty directory
 #	Since scoop installs a .config file already, the below commands merge both directories
 cd ~
 mv .config .config~
 [console]::beep(180,500)  # Audio signal for input
 git clone git@github.com:$user/.config.git
-mv .config~/* .config
-rm -r .config~
-cd ~/docs
+mv .config~/scoop .config
+rm -r -force .config~
+cd $usb
+if ( -not (Test-Path -Path '~/.config/vim' -PathType Container) ) { [Environment]::SetEnvironmentVariable('SCOOP', $env:SCOOP, 'User') ; read-host ABORT: config not loaded ; exit }
 
 echo ''
 echo '__________________________________________________'
@@ -79,7 +76,14 @@ cp ~/.config/mintty/minttyrc ~/apps/scoop/apps/git/current/etc
 
 echo ''
 echo '__________________________________________________'
+echo 'Clone project files'
+cd ~/docs
+~/.config/windows-setup/clone-repos.ps1
+
+echo ''
+echo '__________________________________________________'
 echo 'Install other tools'
+scoop install gh
 scoop bucket add extras
 scoop install mupdf sharex
